Delete __pycache__/ after unit test is done

+++ python/test/test_application.py
@@ -1,15 +1,21 @@
 import unittest
 import os
 import shutil
+import glob
 
 class TestApplication(unittest.TestCase):
     def setUp(self):
-        self.dirname = os.path.join('test', 'tmp')
+        self.dirname  = os.path.join('test', 'tmp')
+        self.pycaches = glob.glob(os.path.join('.', '**', '__pycache__'))
         if not os.path.exists(self.dirname):
             os.makedirs(self.dirname)
 
     def tearDown(self):
-        shutil.rmtree(self.dirname)
+        if os.path.isdir(self.dirname):
+            shutil.rmtree(self.dirname)
+        for pycache in self.pycaches:
+            if os.path.isdir(pycache):
+                shutil.rmtree(pycache)
 
 if __name__ == '__main__':
     unittest.main()
